% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CITwoSide.R
\name{CITwoSide}
\alias{CITwoSide}
\alias{citwoside}
\title{Finding Asymptotically Unbiased Two-Sided Confidence Intervals}
\usage{
CITwoSide(alpha, Z, size, b, q)
}
\arguments{
\item{alpha}{The confidence level of the returned confidence interval}

\item{Z}{A Binomial sample with Tulap noise}

\item{size}{The number of trials in Binomial distribution (parameter n in
Binomial(n, \eqn{\theta}))}

\item{b}{The discrete Laplace noise parameter, obtained by
\eqn{exp(-\epsilon)}}

\item{q}{The truncated quantiles}
}
\value{
An asymptotically unbiased two-sided confindence interval
}
\description{
Finding Asymptotically Unbiased Two-Sided Confidence Intervals
}
\examples{
CITwoSide(alpha = 0.05, Z = 18, size = 30, b = exp(-1), q = 0.05) #(0.411, 0.767)

}
\references{
Awan, Jordan Alexander, and Aleksandra Slavkovic. 2020.
  "Differentially Private Inference for Binomial Data". Journal of Privacy
  and Confidentiality 10 (1). \url{https://doi.org/10.29012/jpc.725}.
}
\seealso{
Finding one-sided confidence intervals \code{\link{CILower}} and
  \code{\link{CIUpper}}
}
