% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpPropTest.R
\name{dpPropTest}
\alias{dpPropTest}
\title{DP Population Proportion Test}
\usage{
dpPropTest(Z, size, theta, b, q, alternative, alpha)
}
\arguments{
\item{Z}{A Binomial sample with Tulap noise}

\item{size}{The number of trials in Binomial distribution (parameter n in
Binomial(n, \eqn{\theta}))}

\item{theta}{The success probability for each trial (parameter \eqn{\theta}
in Binomial(n, \eqn{\theta}))}

\item{b}{Discrete Laplace noise parameters, obtained by \eqn{exp(-\epsilon)}}

\item{q}{The truncated quantiles}

\item{alternative}{The type of alternative hypothesis, must be
\code{"two.sided"}, \code{"less"}, or \code{"greater"}}

\item{alpha}{The confidence level of the confidence interval}
}
\value{
A table summary of the sample estimate, the p-value of the test, and
  the confidence interval of \eqn{\theta}
}
\description{
DP Population Proportion Test
}
\examples{
dpPropTest(9, 30, 0.5, exp(-1), 0.05, "less", 0.05)
}
