% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMP.R
\name{UMP}
\alias{UMP}
\alias{ump}
\alias{umpleft}
\alias{umpright}
\alias{umpLeft}
\alias{umpRight}
\title{Calculating Simple and One-Sided DP-UMP Tests}
\usage{
umpLeft(theta, size, alpha, epsilon, delta)

umpRight(theta, size, alpha, epsilon, delta)
}
\arguments{
\item{theta}{The success probability for each trial (parameter \eqn{\theta}
in Binomial(n, \eqn{\theta}))}

\item{size}{The number of trials in Binomial distribution (parameter n in
Binomial(n, \eqn{\theta}))}

\item{alpha}{Level of the tests}

\item{epsilon}{Parameter \eqn{\epsilon} in \eqn{(\epsilon, \delta)}-DP}

\item{delta}{Parameter \eqn{\delta} in \eqn{(\epsilon, \delta)}-DP}
}
\value{
A vector of one-sided DP-UMP tests
}
\description{
Calculating Simple and One-Sided DP-UMP Tests
}
\examples{
#leftUMP
left <- umpLeft(theta = 0.4, size = 10, alpha = 0.05, epsilon = 1, delta = 0.01)

#rightUMP
right <- umpRight(theta = 0.4, size = 10, alpha = 0.05, epsilon = 1, delta = 0.01)

#plot
plot(left, type = "l", main = "One-sided DP UMP",
     xlab = "x", ylab = "Phi (Probability of rejecting the null hypothesis)")
lines(right, lty = 2, col = "blue")
legend("left", legend=c("Left UMP", "Right UMP"), col=c("black", "blue"), lty=1:2, lwd = 1.5)

}
\references{
Awan, Jordan Alexander, and Aleksandra Slavkovic. 2020.
  "Differentially Private Inference for Binomial Data". Journal of Privacy
  and Confidentiality 10 (1). \url{https://doi.org/10.29012/jpc.725}.
}
\seealso{
Calculating unbiased two-sided DP-UMP tests (\code{\link{UMPU}}) and
  asymptotically unbiased two-sided DP-UMP tests (\code{\link{umpuApprox}})
}
