% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umpuApprox.R
\name{umpuApprox}
\alias{umpuApprox}
\alias{umpuapprox}
\title{Asymptotically Unbiased Two-Sided DP-UMP Tests}
\usage{
umpuApprox(theta, size, alpha, epsilon, delta)
}
\arguments{
\item{theta}{The success probability for each trial (parameter \eqn{\theta}
in Binomial(n, \eqn{\theta}))}

\item{size}{The number of trials in Binomial distribution (parameter n in
Binomial(n, \eqn{\theta}))}

\item{alpha}{Level of the tests}

\item{epsilon}{Parameter \eqn{\epsilon} in \eqn{(\epsilon, \delta)}-DP}

\item{delta}{Parameter \eqn{\delta} in \eqn{(\epsilon, \delta)}-DP}
}
\value{
A vector of asymptotically unbiased two-sided DP-UMP tests
}
\description{
Asymptotically Unbiased Two-Sided DP-UMP Tests
}
\examples{
#Comparing unbiased DP-UMP tests obtained by umpuApprox and UMPU
asymp <- umpuApprox(theta = 0.4, size = 10, alpha = 0.05, epsilon = 1, delta = 0.01)
twoside <- UMPU(theta = 0.4, size = 10, alpha = 0.05, epsilon = 1, delta = 0.01)

#Plot the probability of rejecting the null hypothesis based on x
plot(asymp, type = "l", lwd = 1.5, col = "red", xlab = "x",
  ylab = "Phi (Probability of rejecting the null hypothesis)",
  main = "Unbiased DP-UMP Tests")
lines(twoside, type = "l", lwd = 1.5, lty = 2, col = "blue")
legend("topleft", legend=c("Asymptotically UMPU", "UMPU"),
  col=c("red", "blue"), lty=1:2, lwd = 1.5)

}
\references{
Awan, Jordan Alexander, and Aleksandra Slavkovic. 2020.
  "Differentially Private Inference for Binomial Data". Journal of Privacy
  and Confidentiality 10 (1). \url{https://doi.org/10.29012/jpc.725}.
}
\seealso{
Calculating simple and one-sided DP-UMP tests (\code{\link{umpLeft}}
  or \code{\link{umpRight}}) and unbiased two-sided DP-UMP tests
  (\code{\link{UMPU}})
}
