% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIOneSide.R
\name{CIOneSide}
\alias{CIOneSide}
\alias{cioneside}
\alias{cileft}
\alias{ciright}
\alias{CILeft}
\alias{CIRight}
\title{Finding One-Sided Confidence Intervals}
\usage{
CILeft(alpha, Z, size, b, q)

CIRight(alpha, Z, size, b, q)
}
\arguments{
\item{alpha}{The level of confidence intervals}

\item{Z}{A Binomial sample with Tulap noise}

\item{size}{The number of trials in Binomial distribution (parameter n in
Binomial(n, \eqn{\theta}))}

\item{b}{Discrete Laplace noise parameters, obtained by \eqn{exp(-\epsilon)}}

\item{q}{The truncated quantiles}
}
\value{
A lower bound
}
\description{
Finding One-Sided Confidence Intervals
}
\examples{
CILeft(0.05, 7, 10, exp(-1), 0.05) #confidence interval (0.384, 1)
CIRight(0.05, 2, 10, exp(-1), 0.05) #confidence interval (0, 0.517)
}
\references{
Awan, Jordan Alexander, and Aleksandra Slavkovic. 2020.
  "Differentially Private Inference for Binomial Data". Journal of Privacy
  and Confidentiality 10 (1). \url{https://doi.org/10.29012/jpc.725}.
}
\seealso{
Finding asymptotically unbiased two-sided confidence intervals
  \code{\link{CITwoSide}}
}
